test-automation-framework/requirements.txt
# Core testing framework
behave==1.2.6
behave-html-formatter==0.9.10
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1

# UI Testing
playwright==1.40.0
selenium==4.15.2

# API Testing
requests==2.31.0
jsonschema==4.19.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
faker==20.1.0
allure-behave==2.13.2

# Logging and reporting
colorlog==6.8.0
jinja2==3.1.2

# Data handling
pandas==2.1.3
openpyxl==3.1.2

# Parallel execution
pytest-parallel==0.1.1